+++ release_notes.md
@@ -11,7 +11,7 @@
   - [#232](https://github.com/Mehdik/Humanizer/pull/232): Adding code & tests to handle Arabic numbers to ordinal 
   - [#235](https://github.com/Mehdik/Humanizer/pull/235): Fixed the conversion for "1 millon" in SpanishNumberToWordsConverter
   - [#233](https://github.com/Mehdik/Humanizer/pull/233): Added build.cmd and Verify build configuration for strict project build and analysis
-  - [#227](https://github.com/MehdiK/Humanizer/pull/227): Switched Configurator.FormatterFactories to have a public get to allow for adding/removing of formatters
+  - [#227](https://github.com/MehdiK/Humanizer/pull/227): Moved factory collections to their own classes, allowed public access via Configurator, and made the default factories lazy loaded
 
 [Commits](https://github.com/MehdiK/Humanizer/compare/v1.23.1...v1.24.1)